requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyjwt==2.9.0
passlib[argon2]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
